+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.baseballgame.Controller.BaseBallGameController;
+import baseball.controller.BaseBallGameController;
 
 public class Application {
     public static void main(String[] args) {
