+++ src/main/java/baseball/view/inputview/InputViewValidator.java
@@ -1,4 +1,4 @@
-package baseball.inputview;
+package baseball.view.inputview;
 
 import java.util.Arrays;
