+++ src/test/java/baseball/baseballgame/GameOverCheckerTest.java
@@ -2,6 +2,7 @@ package baseball.baseballgame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.model.GameOverChecker;
 import java.util.List;
 import org.junit.jupiter.api.Test;
