+++ src/test/java/baseball/baseballgame/ball/BallNumberTest.java
@@ -1,10 +1,11 @@
 package baseball.baseballgame.ball;
 
-import static baseball.baseballgame.ball.BallNumber.BALL_NUMBER_ERROR;
+import static baseball.model.ball.BallNumber.BALL_NUMBER_ERROR;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import baseball.model.ball.BallNumber;
 import org.junit.jupiter.api.Test;
 
 public class BallNumberTest {
