+++ src/test/java/baseball/baseballgame/balls/BallReportTest.java
@@ -2,6 +2,8 @@ package baseball.baseballgame.balls;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.model.balls.BallReport;
+import baseball.model.balls.JudgementType;
 import java.util.List;
 import org.junit.jupiter.api.Test;
