+++ src/test/java/baseball/baseballgame/balls/BallsTest.java
@@ -1,11 +1,14 @@
 package baseball.baseballgame.balls;
 
-import static baseball.baseballgame.balls.Balls.BALLS_RANGE_ERROR;
+import static baseball.model.balls.Balls.BALLS_RANGE_ERROR;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import baseball.baseballgame.ball.Ball;
-import baseball.baseballgame.ball.BallNumber;
+import baseball.model.ball.Ball;
+import baseball.model.ball.BallNumber;
+import baseball.model.balls.BallReport;
+import baseball.model.balls.Balls;
+import baseball.model.balls.JudgementType;
 import java.util.List;
 import org.junit.jupiter.api.Test;
